+++ src/ui/dyn_content/mod.rs
@@ -5,7 +5,7 @@ use super::util;
 use super::MainWindow;
 use super::ResultsDynType;
 use super::DYN_CONTENT_VIEW;
-use indexmap::IndexMap;
+use dashmap::DashMap;
 use slint::Weak;
 use std::sync::Arc;
 use tokio::sync::Mutex;
@@ -30,7 +30,7 @@ use tokio::sync::Mutex;
 pub fn setup_content_display<E, F>(
     ui: Weak<MainWindow>,
     results: Arc<Mutex<Option<ResultsDynType<E, F>>>>,
-    results_awaited: Arc<Mutex<IndexMap<String, Vec<String>>>>,
+    results_awaited: Arc<DashMap<String, Vec<String>>>,
     index: Arc<Mutex<usize>>,
     content_index: Arc<Mutex<usize>>,
     results_type: ResultType,
@@ -96,7 +96,7 @@ pub fn setup_content_display<E, F>(
 fn get_resource_enter_fn<E, F>(
     ui: Weak<MainWindow>,
     results: Arc<Mutex<Option<ResultsDynType<E, F>>>>,
-    results_awaited: Arc<Mutex<IndexMap<String, Vec<String>>>>,
+    results_awaited: Arc<DashMap<String, Vec<String>>>,
     index: Arc<Mutex<usize>>,
     content_index: Arc<Mutex<usize>>,
     results_type: ResultType,
@@ -122,15 +122,13 @@ where
             let locked = futures::join!(
                 results_clone.lock(),
                 index_clone.lock(),
-                results_awaited_clone.lock(),
             );
             let mut results_lock = locked.0;
             let index_lock = locked.1;
-            let mut results_awaited_lock = locked.2;
 
             let content = match results_lock.as_mut() {
                 Some(results) => match results {
-                    Ok(results) => match results.get_index_mut(*index_lock) {
+                    Ok(results) => match results.get_mut(*index_lock) {
                         Some(result) => {
                             // show the view dynamic content window
                             let ui_clone = ui.clone();
@@ -143,10 +141,11 @@ where
                                 util::slint_event_loop_panic(err);
                             };
                             // get content
-                            match results_awaited_lock.get(result.0) {
+                            match results_awaited_clone.get(&result.0) {
                                 Some(result) => result,
                                 None => {
-                                    let awaited = match result.1.await {
+                                    let (title, handle) = result;
+                                    let awaited = match handle.await {
                                         Ok(handled) => match handled {
                                             Ok(content) => content,
                                             Err(error) => {
@@ -164,10 +163,10 @@ where
                                     };
 
                                     // save already awaited
-                                    results_awaited_lock.insert(result.0.to_owned(), awaited);
+                                    results_awaited_clone.insert(title.to_owned(), awaited);
 
                                     // unwrap is safe since we just inserted this element
-                                    results_awaited_lock.get(result.0).unwrap()
+                                    results_awaited_clone.get(title).unwrap()
                                 }
                             }
                         }
@@ -198,7 +197,6 @@ where
             let first = content.first().unwrap().to_owned();
             // drop the Mutex locks
             drop(results_lock);
-            drop(results_awaited_lock);
             drop(index_lock);
             if let Err(err) = slint::invoke_from_event_loop(move || {
                 let ui_strong = util::get_ui(ui_clone);